mcp-servers/python/granite_vision_server/pyproject.toml
[project]
name = "granite-vision-server"
version = "2.0.0"
description = "Create a comprehensive MCP Server for IBM Granite Vision Models supporting document understanding, image analysis, OCR, visual question answering, and multi-modal processing with enterprise deployment capabilities."
authors = [
  { name = "Anna Topol, Łukasz Strąk, Hong Wei Jia, Lisette Contreras, Mohammed Kazmi", email = "noreply@example.com" }
]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastmcp>=2.11.3",
  "pydantic>=2.5.0",
  "ollama>=0.4.0",
  "Pillow>=10.0.0",
  "requests>=2.31.0",
  "typing-extensions>=4.5.0",
  "pdf2image>=1.17.0",
  "pypdf>=6.4.0",
  "pandas>=2.3.3",
  "pytesseract>=0.3.13",
  "transformers>=4.57.3",
  "torch>=2.9.1",
  "ruff>=0.14.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-asyncio>=0.21.0",
  "pytest-cov>=4.0.0",
  "black>=23.0.0",
  "mypy>=1.5.0",
  "ruff>=0.0.290",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/granite_vision_server"]

[project.scripts]
granite-vision-server = "granite_vision_server.server_fastmcp:main"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "W", "F", "B", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=granite_vision_server --cov-report=term-missing"
